swap example to typescript lib that will include the binary
